In Password Manager project, ask for user confirmation before adding to password file.

+++ Password Manager GUI/main.py
@@ -1,4 +1,5 @@
 from tkinter import *
+from tkinter import messagebox
 from PIL import Image, ImageTk
 
 DEFAULT_EMAIL = "dennis@example.com"
@@ -12,11 +13,16 @@ def save_pwd():
     str_username = ent_username.get()
     str_password = ent_password.get()
 
-    with open("data.txt", mode='w') as file:
-        file.writelines(f"{str_website} | {str_username} | {str_password}")
-        ent_website.delete(0, END)
-        ent_username.delete(0, END)
-        ent_password.delete(0, END)
+    is_ok = messagebox.askokcancel(title="Confirmation", message=f"Adding info to password file:\n\n"
+                                                                 f"Website: \t{str_website}\n"
+                                                                 f"Email/Username: \t{str_username}\n"
+                                                                 f"Password: \t{str_password}")
+    if is_ok:
+        with open("data.txt", mode='w') as file:
+            file.writelines(f"{str_website} | {str_username} | {str_password}")
+            ent_website.delete(0, END)
+            ent_username.delete(0, END)
+            ent_password.delete(0, END)
 # ---------------------------- UI SETUP ------------------------------- #
